Move MetadataDebugSummary to a new file
Also switch from cascade of 'if's to a 'switch' when looking at debug entries
